Fix Validation

+++ BetSystem/Endpoint/BetOnEventEndpoint.cs
@@ -1,5 +1,6 @@
 ﻿using BetSystem.Contract;
 using BetSystem.Contract.BusinnessLogic;
+using BetSystem.Model;
 using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 
@@ -38,12 +39,21 @@ namespace BetSystem.Endpoint
 
         private static IResult DeleteById(int id, [FromServices] IBetOnEventService service)
         {
+            if (service.GetBetOnEventById(id) == null) return Results.NotFound();
+
             service.DeleteBetOnEvent(id);
             return Results.Ok(id);
         }
 
-        private static IResult PutBet(int id, BetOnEventDto betOnEventDto, [FromServices] IBetOnEventService service)
+        private static IResult PutBet(int id, BetOnEventDto betOnEventDto, IValidator<BetOnEventDto> validator, [FromServices] IBetOnEventService service)
         {
+            if (service.GetBetOnEventById(id) == null) return Results.NotFound();
+            var result = validator.Validate(betOnEventDto);
+            if (!result.IsValid)
+            {
+                return Results.ValidationProblem(result.ToDictionary());
+            }
+
             service.PutBetOnEvent(id, betOnEventDto);
             return Results.NoContent();
         }
